Fold `IntotoPayload` into `SigningPayload`.
We want to support only one signing payload type.
