+++ src/model_signing/sign.py
@@ -27,7 +27,6 @@ from typing import Optional
 
 from model_signing import hash
 from model_signing import manifest
-from model_signing.signing import in_toto
 from model_signing.signing import sign_sigstore as sigstore
 from model_signing.signing import signing
 
@@ -61,7 +60,7 @@ class Config:
     def __init__(self):
         """Initializes the default configuration for signing."""
         self._hashing_config = hash.Config()
-        self._payload_generator = in_toto.IntotoPayload.from_manifest
+        self._payload_generator = signing.SigningPayload.from_manifest
         self._signer = sigstore.SigstoreSigner(
             use_ambient_credentials=False, use_staging=False
         )
