+++ tests/signing/testdata/signing/TestSigningPayload/empty_model_file
