+++ tests/signing/testdata/signing/TestSigningPayload/model_folder_with_empty_file
