+++ tests/signing/testdata/signing/TestSigningPayload/sample_model_file
