+++ tests/signing/testdata/signing/TestSigningPayload/symlink_model_folder
